requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-genai>=1.50.0
python-dotenv==1.0.1
requests==2.32.0
python-multipart==0.0.9

